# Conflicts:
#	src/main/java/com/comp5703/Neighbourhood/Walk/Entities/Users.java

+++ src/main/java/com/comp5703/Neighbourhood/Walk/Repository/UsersRepository.java
@@ -4,6 +4,6 @@ import com.comp5703.Neighbourhood.Walk.Entities.Users;
 import org.springframework.data.repository.CrudRepository;
 import java.util.Optional;
 
-public interface UsersRepository extends CrudRepository<Users, Integer> {
+public interface UsersRepository extends CrudRepository<Users, Long> {
     Optional<Users> findByEmail(String email);
 }
