+++ src/main/java/com/comp5703/Neighbourhood/Walk/Service/Impl/UsersServiceImpl.java
@@ -25,7 +25,7 @@ public class UsersServiceImpl implements UsersService {
     }
 
     @Override
-    public void deleteUsers(int id) {
+    public void deleteUsers(long id) {
         usersRepository.deleteById(id);
     }
